+++ AlphaFS/Win32Errors.cs
@@ -64,7 +64,10 @@ namespace Alphaleonis.Win32
       //public const uint ERROR_BAD_ENVIRONMENT = 10;
       //public const uint ERROR_BAD_FORMAT = 11;
       //public const uint ERROR_INVALID_ACCESS = 12;
-      //public const uint ERROR_INVALID_DATA = 13;
+
+      /// <summary>(13) The data is invalid.</summary>
+      public const uint ERROR_INVALID_DATA = 13;
+
       //public const uint ERROR_OUTOFMEMORY = 14;
 
       /// <summary>(15) The system cannot find the drive specified.</summary>
@@ -294,7 +297,7 @@ namespace Alphaleonis.Win32
       //public const uint WAIT_TIMEOUT = 258;   
 
       ///// <summary>(259) No more data is available.</summary>
-      //public const uint ERROR_NO_MORE_ITEMS = 259;
+      public const uint ERROR_NO_MORE_ITEMS = 259;
 
       //public const uint ERROR_CANNOT_COPY = 266;